Restructure fluent classes
