+++ coupling_components/solver_wrappers/fluent/v2019R2.py
@@ -1,12 +1,13 @@
-from coconut.coupling_components.solver_wrappers.fluent.v2019R1 import SolverWrapperFluent2019R1
+from coconut.coupling_components.solver_wrappers.fluent.fluent import SolverWrapperFluent
 
 
 def create(parameters):
     return SolverWrapperFluent2019R2(parameters)
 
 
-class SolverWrapperFluent2019R2(SolverWrapperFluent2019R1):
+class SolverWrapperFluent2019R2(SolverWrapperFluent):
+    version = '2019R2'
+    version_bis = '19.4.0'
 
-    def set_fluent_version(self):
-        self.version = '2019R2'
-        self.version_bis = '19.4.0'
+    def __init__(self, parameters):
+        super().__init__(parameters)
