+++ coupling_components/solver_wrappers/fluent/v2019R3.py
@@ -1,12 +1,13 @@
-from coconut.coupling_components.solver_wrappers.fluent.v2019R2 import SolverWrapperFluent2019R2
+from coconut.coupling_components.solver_wrappers.fluent.fluent import SolverWrapperFluent
 
 
 def create(parameters):
     return SolverWrapperFluent2019R3(parameters)
 
 
-class SolverWrapperFluent2019R3(SolverWrapperFluent2019R2):
+class SolverWrapperFluent2019R3(SolverWrapperFluent):
+    version = '2019R3'
+    version_bis = '19.5.0'
 
-    def set_fluent_version(self):
-        self.version = '2019R3'
-        self.version_bis = '19.5.0'
+    def __init__(self, parameters):
+        super().__init__(parameters)
