+++ tests/solver_wrappers/fluent/test_v2019R2.py
@@ -1,15 +1,14 @@
-from coconut.tests.solver_wrappers.fluent.test_v2019R1 \
-    import TestSolverWrapperFluent2019R1Tube2D, TestSolverWrapperFluent2019R1Tube3D
+from coconut.tests.solver_wrappers.fluent.fluent import TestSolverWrapperFluentTube2D, TestSolverWrapperFluentTube3D
 
 import unittest
 
 
-class TestSolverWrapperFluent2019R2Tube2D(TestSolverWrapperFluent2019R1Tube2D):
+class TestSolverWrapperFluent2019R2Tube2D(TestSolverWrapperFluentTube2D):
     version = '2019R2'
     setup_case = True
 
 
-class TestSolverWrapperFluent2019R2Tube3D(TestSolverWrapperFluent2019R1Tube3D):
+class TestSolverWrapperFluent2019R2Tube3D(TestSolverWrapperFluentTube3D):
     version = '2019R2'
     setup_case = True
